Adding initial SCR and LBR files
These are not complete as is. We still need to decide on exact project
examples (e.g. Bare Bones Arduino, Simon PTH). And then we need to make
sure SPK1.lbr has all the necessary devices to create these projects.
SPK.SCR is pretty close. It is very similar to the standard SFE eagle
script file, but it also clears the used libraries and only uses the
SPK1.lbr